refactor: splitted New Request Form implementation in multiple files
