+++ src/modules/newRequestForm/NewRequestForm.tsx
@@ -1,0 +1,14 @@
+import { TicketFormField } from "./TicketFormField";
+import { TicketForm } from "./types";
+
+export interface NewRequestFormProps {
+  ticketForms: TicketForm[];
+}
+
+export function NewRequestForm({ ticketForms }: NewRequestFormProps) {
+  return (
+    <form>
+      <TicketFormField ticketForms={ticketForms} />
+    </form>
+  );
+}
