+++ src/modules/newRequestForm/TicketFormField.tsx
@@ -1,0 +1,33 @@
+import {
+  Combobox,
+  Field,
+  IComboboxProps,
+  Label,
+  Option,
+} from "@zendeskgarden/react-dropdowns.next";
+import { TicketForm } from "./types";
+
+interface TicketFormFieldProps {
+  ticketForms: TicketForm[];
+}
+
+export function TicketFormField({ ticketForms }: TicketFormFieldProps) {
+  const handleChange: IComboboxProps["onChange"] = ({ selectionValue }) => {
+    if (selectionValue && typeof selectionValue === "string") {
+      window.location.href = selectionValue;
+    }
+  };
+
+  return (
+    <Field>
+      <Label>Please choose your issue below</Label>
+      <Combobox isEditable={false} onChange={handleChange}>
+        {ticketForms.map(({ id, url, name }) => (
+          <Option key={id} value={url} label={name}>
+            {name}
+          </Option>
+        ))}
+      </Combobox>
+    </Field>
+  );
+}
