+++ src/modules/newRequestForm/types.tsx
@@ -1,0 +1,5 @@
+export interface TicketForm {
+  id: string;
+  url: string;
+  name: string;
+}
