+++ templates/document_head.hbs
@@ -7,7 +7,7 @@
         "imports": {
             "vendor": "{{asset 'vendor.js'}}",
             "shared": "{{asset 'shared.js'}}",
-            "new-request-form": "{{asset 'NewRequestForm.js'}}"
+            "new-request-form": "{{asset 'newRequestForm.js'}}"
         }
     }
 </script>
